@@ -4,8 +4,6 @@ import { CommonModule } from '@angular/common';
 import { DashboardRoutingModule } from './dashboard-routing.module';
 import { BaseComponent } from './base/base.component';
 import { CarouselComponent } from './base/carousel/carousel.component';
-import { NavbarComponent } from './base/navbar/navbar.component';
-import { MenubarModule } from 'primeng/menubar';
 import { InputTextModule } from 'primeng/inputtext';
 import { ButtonModule } from 'primeng/button';
 import { HttpClientModule } from '@angular/common/http';
@@ -16,19 +14,24 @@ import {RatingModule} from 'primeng/rating';
 import {DataViewModule} from 'primeng/dataview';
 import { AppService } from '../../core/services/app.service';
 import { ProductsViewComponent } from './base/products-view/products-view.component';
+import { ProductComponent } from './base/product/product.component';
+import {GalleriaModule} from 'primeng/galleria';
+import { PhotoService } from 'src/core/services/photo.service';
+import { DividerModule } from 'primeng/divider';
+import {TableModule} from 'primeng/table';
+
 
 @NgModule({
   declarations: [
     BaseComponent,
     CarouselComponent,
-    NavbarComponent,
     ProductsViewComponent,
+    ProductComponent,
 
   ],
   imports: [
     CommonModule,
     DashboardRoutingModule,
-    MenubarModule,
     InputTextModule,
     ButtonModule,
     CarouselModule,
@@ -37,8 +40,12 @@ import { ProductsViewComponent } from './base/products-view/products-view.compon
     DataViewModule,
     DropdownModule,
     RatingModule,
-    FormsModule
+    FormsModule,
+    GalleriaModule,
+    DividerModule,
+    TableModule
+    
   ],
-  providers:[HttpClientModule,AppService]
+  providers:[HttpClientModule,AppService,PhotoService]
 })
 export class DashboardModule { }
