@@ -1,7 +1,9 @@
-import { Component } from '@angular/core';
+import { Component, ElementRef, HostListener, Input } from '@angular/core';
 import { AppService } from '../../../../core/services/app.service';
-import { Product } from '../../../../core/models/model';
+import { Media, Product } from '../../../../core/models/model';
 import {SelectItem} from 'primeng/api';
+import { environment } from 'src/environments/environment.development';
+import { Router } from '@angular/router';
 
 
 @Component({
@@ -10,6 +12,7 @@ import {SelectItem} from 'primeng/api';
   styleUrls: ['./products-view.component.scss']
 })
 export class ProductsViewComponent {
+  @Input() appImgFallback:string;
 
   products: Product[];
 
@@ -21,11 +24,19 @@ export class ProductsViewComponent {
 
     sortKey:string;
 
-    constructor(private appServices: AppService) { }
+    mediaUrl:string
+
+    constructor(private appServices: AppService,private router:Router) { 
+      this.mediaUrl=environment.MEDIA_API_URL
+      console.log(this.mediaUrl)
+    }
 
     ngOnInit() {
-        this.appServices.get("assets/products.json").subscribe(
-          res => this.products = res.data
+        this.appServices.get("inventory/products/").subscribe(
+          res =>{
+            console.log(res.data.results)
+            this.products = res.data.results
+          }
         );
         this.sortOptions = [
             {label: 'Price High to Low', value: '!price'},
@@ -44,5 +55,18 @@ export class ProductsViewComponent {
             this.sortOrder = 1;
             this.sortField = value;
         }
+    }
+
+    productPage(id:string){
+      this.router.navigate(['product/',id])
+    }
+
+    getMediaUrl(product:any){
+      console.log(product?.media[0]!=undefined);
+      if (product.media[0] !=undefined){
+        return this.mediaUrl+ product.media[0].img_url;
+      }else{
+        return this.mediaUrl+'/images/default.png'
+      }
     }
 }
