@@ -1,12 +1,46 @@
+export interface Media {
+    
+    img_url?: string,
+    alt_text?: string,
+    is_feature?: boolean
+    
+}
 export interface Product{
-    _id?: string;
-    name?: string;
-    image?: string;
-    description?:string;
-    brand?: string;
-    category?: string;
-    price?: number;
-    countInStock?: number;
-    rating?: number;
-    numReviews?: number;
+    id?: number,
+    sku?: string,
+    store_price?: string,
+    is_default?: boolean,
+    brand?: {
+        name?: string
+    },
+    product?: {
+        name: string,
+        web_id : string,
+        category?:{
+            name: string,
+            slug: string,
+            is_active: boolean
+        },
+        description?:string,
+        rating:"string",
+        no_reviews:number
+    },
+    weight?: number,
+    media?: Media[],
+    attributes?: [
+        {
+            attribute_value?: string,
+            product_attribute?: {
+                id?: number,
+                name?: string,
+                description?: string
+            }
+        }
+    ],
+    product_type?: {
+        name? : string
+    }
+    sale_price: string,
+    units: number,
+    units_sold: number
 }
